requirements.txt
urllib3==1.26.9
requests==2.28.1
Pillow==9.1.1
google==3.0.0
google-api-python-client==2.52.0
google-auth-oauthlib==0.5.2